Revert "Add non-minified geofire-utils.js build target"
