+++ .github/workflows/test.yaml
@@ -27,7 +27,6 @@ jobs:
           key: ${{ runner.os }}-node-${{ matrix.node-version }}-${{ hashFiles('**/package-lock.json') }}
 
       - run: npm install
-      - run: npm run bootstrap
       - run: npm run lint
       - run: npm run test
       - run: npm run coverage
@@ -36,4 +35,3 @@ jobs:
         uses: coverallsapp/github-action@master
         with:
           github-token: ${{ secrets.GITHUB_TOKEN }}
-          path-to-lcov: packages/geofire/coverage/lcov.info
