--- lerna.json
@@ -1,6 +1,0 @@
-{
-  "version": "5.0.1",
-  "packages": [
-    "packages/*"
-  ]
-}
