--- packages/geofire-common/test/Main.test.ts
@@ -1,9 +1,0 @@
-import * as chai from 'chai';
-
-const expect = chai.expect;
-
-describe('GeoFire GeoCallbackRegistration Tests:', () => {
-  it("hello world", () => {
-    expect(true).to.be.true;
-  });
-});
