--- packages/geofire-common/test/mocha.opts
@@ -1,5 +1,0 @@
---reporter spec
---require ts-node/register
---require jsdom-global/register
---timeout 30000
---exit test/**/*.test.ts
