--- packages/geofire-common/tsconfig.json
@@ -1,22 +1,0 @@
-{
-  "compileOnSave": false,
-  "compilerOptions": {
-    "declaration": true,
-    "importHelpers": true,
-    "lib": [
-      "es2015",
-      "dom"
-    ],
-    "moduleResolution": "node",
-    "noImplicitAny": false,
-    "outDir": "dist",
-    "sourceMap": false,
-    "target": "es5",
-    "typeRoots": [
-      "node_modules/@types"
-    ]
-  },
-  "include": [
-    "src"
-  ]
-}
