+++ src/GeoCallbackRegistration.ts
