+++ src/GeoQuery.ts
@@ -1,64 +1,8 @@
 import { GeoCallbackRegistration } from './GeoCallbackRegistration';
 import {
-  distance, encodeGeohash, geohashQueries, validateLocation
-} from 'geofire-common';
-import {
-  decodeGeoFireObject, geoFireGetKey } from './databaseUtils';
-import * as GeoFireTypes from './GeoFireTypes';
-import * as DatabaseTypes from '@firebase/database-types';
-
-export interface QueryCriteria {
-  center?: number[];
-  radius?: number;
-}
-export type QueryStateCallback = (
-  a: DatabaseTypes.DataSnapshot | null,
-  b?: string
-) => any;
-export interface QueryState {
-  active: boolean;
-  childAddedCallback: QueryStateCallback;
-  childRemovedCallback: QueryStateCallback;
-  childChangedCallback: QueryStateCallback;
-  valueCallback: QueryStateCallback;
-}
-/**
- * Validates the inputted query criteria and throws an error if it is invalid.
- *
- * @param newQueryCriteria The criteria which specifies the query's center and/or radius.
- * @param requireCenterAndRadius The criteria which center and radius required.
- */
-export function validateCriteria(newQueryCriteria: QueryCriteria, requireCenterAndRadius = false): void {
-  if (typeof newQueryCriteria !== 'object') {
-    throw new Error('query criteria must be an object');
-  } else if (typeof newQueryCriteria.center === 'undefined' && typeof newQueryCriteria.radius === 'undefined') {
-    throw new Error('radius and/or center must be specified');
-  } else if (requireCenterAndRadius && (typeof newQueryCriteria.center === 'undefined' || typeof newQueryCriteria.radius === 'undefined')) {
-    throw new Error('query criteria for a new query must contain both a center and a radius');
-  }
-
-  // Throw an error if there are any extraneous attributes
-  const keys: string[] = Object.keys(newQueryCriteria);
-  for (const key of keys) {
-    if (key !== 'center' && key !== 'radius') {
-      throw new Error('Unexpected attribute \'' + key + '\' found in query criteria');
-    }
-  }
-
-  // Validate the 'center' attribute
-  if (typeof newQueryCriteria.center !== 'undefined') {
-    validateLocation(newQueryCriteria.center);
-  }
-
-  // Validate the 'radius' attribute
-  if (typeof newQueryCriteria.radius !== 'undefined') {
-    if (typeof newQueryCriteria.radius !== 'number' || isNaN(newQueryCriteria.radius)) {
-      throw new Error('radius must be a number');
-    } else if (newQueryCriteria.radius < 0) {
-      throw new Error('radius must be greater than or equal to 0');
-    }
-  }
-}
+  decodeGeoFireObject, distance, encodeGeohash, geoFireGetKey, geohashQueries, validateCriteria, validateLocation
+} from './utils';
+import { GeoFireTypes } from './GeoFireTypes';
 
 /**
  * Creates a GeoQuery instance.
@@ -70,7 +14,7 @@ export class GeoQuery {
   private _cancelled = false;
   private _center: number[];
   // A dictionary of geohash queries which currently have an active callbacks
-  private _currentGeohashesQueried: { [name: string]: QueryState } = {};
+  private _currentGeohashesQueried: { [name: string]: GeoFireTypes.QueryState } = {};
   // A dictionary of locations that a currently active in the queries
   // Note that not all of these are currently within this query
   private _locationsTracked: { [name: string]: GeoFireTypes.LocationTracked } = {};
@@ -88,7 +32,7 @@ export class GeoQuery {
    * @param _firebaseRef A Firebase reference where the GeoFire data will be stored.
    * @param queryCriteria The criteria which specifies the query's center and radius.
    */
-  constructor(private _firebaseRef: DatabaseTypes.Reference, queryCriteria: QueryCriteria) {
+  constructor(private _firebaseRef: GeoFireTypes.firebase.Reference, queryCriteria: GeoFireTypes.QueryCriteria) {
     // Firebase reference of the GeoFire which created this query
     if (Object.prototype.toString.call(this._firebaseRef) !== '[object Object]') {
       throw new Error('firebaseRef must be an instance of Firebase');
@@ -227,7 +171,7 @@ export class GeoQuery {
    *
    * @param newQueryCriteria The criteria which specifies the query's center and radius.
    */
-  public updateCriteria(newQueryCriteria: QueryCriteria): void {
+  public updateCriteria(newQueryCriteria: GeoFireTypes.QueryCriteria): void {
     // Validate and save the new query criteria
     validateCriteria(newQueryCriteria);
     this._center = newQueryCriteria.center || this._center;
@@ -276,7 +220,7 @@ export class GeoQuery {
    * @param query The geohash query.
    * @param queryState An object storing the current state of the query.
    */
-  private _cancelGeohashQuery(query: string[], queryState: QueryState): void {
+  private _cancelGeohashQuery(query: string[], queryState: GeoFireTypes.QueryState): void {
     const queryRef = this._firebaseRef.orderByChild('g').startAt(query[0]).endAt(query[1]);
     queryRef.off('child_added', queryState.childAddedCallback);
     queryRef.off('child_removed', queryState.childRemovedCallback);
@@ -289,7 +233,7 @@ export class GeoQuery {
    *
    * @param locationDataSnapshot A snapshot of the data stored for this location.
    */
-  private _childAddedCallback(locationDataSnapshot: DatabaseTypes.DataSnapshot): void {
+  private _childAddedCallback(locationDataSnapshot: GeoFireTypes.firebase.DataSnapshot): void {
     this._updateLocation(geoFireGetKey(locationDataSnapshot), decodeGeoFireObject(locationDataSnapshot.val()));
   }
 
@@ -298,7 +242,7 @@ export class GeoQuery {
    *
    * @param locationDataSnapshot A snapshot of the data stored for this location.
    */
-  private _childChangedCallback(locationDataSnapshot: DatabaseTypes.DataSnapshot): void {
+  private _childChangedCallback(locationDataSnapshot: GeoFireTypes.firebase.DataSnapshot): void {
     this._updateLocation(geoFireGetKey(locationDataSnapshot), decodeGeoFireObject(locationDataSnapshot.val()));
   }
 
@@ -307,10 +251,10 @@ export class GeoQuery {
    *
    * @param locationDataSnapshot A snapshot of the data stored for this location.
    */
-  private _childRemovedCallback(locationDataSnapshot: DatabaseTypes.DataSnapshot): void {
+  private _childRemovedCallback(locationDataSnapshot: GeoFireTypes.firebase.DataSnapshot): void {
     const key: string = geoFireGetKey(locationDataSnapshot);
     if (key in this._locationsTracked) {
-      this._firebaseRef.child(key).once('value', (snapshot: DatabaseTypes.DataSnapshot) => {
+      this._firebaseRef.child(key).once('value', (snapshot: GeoFireTypes.firebase.DataSnapshot) => {
         const location: number[] = (snapshot.val() === null) ? null : decodeGeoFireObject(snapshot.val());
         const geohash: string = (location !== null) ? encodeGeohash(location) : null;
         // Only notify observers if key is not part of any other geohash query or this actually might not be
@@ -468,7 +412,7 @@ export class GeoQuery {
       const query: string[] = this._stringToQuery(toQueryStr);
 
       // Create the Firebase query
-      const firebaseQuery: DatabaseTypes.Query = this._firebaseRef.orderByChild('g').startAt(query[0]).endAt(query[1]);
+      const firebaseQuery: GeoFireTypes.firebase.Query = this._firebaseRef.orderByChild('g').startAt(query[0]).endAt(query[1]);
 
       // For every new matching geohash, determine if we should fire the 'key_entered' event
       const childAddedCallback = firebaseQuery.on('child_added', (a) => this._childAddedCallback(a));
