+++ src/index.ts
@@ -1,4 +1,4 @@
 export { GeoCallbackRegistration } from './GeoCallbackRegistration';
 export { GeoFire } from './GeoFire';
 export { GeoQuery } from './GeoQuery';
-export * from './GeoFireTypes';
+export { GeoFireTypes } from './GeoFireTypes';
