+++ test/GeoCallbackRegistration.test.ts
