+++ test/GeoQuery.test.ts
